stage3/pyproject.toml
[project]
name = "mcpserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "mcp[cli]>=1.22.0",
    "psycopg2-binary>=2.9.11",
    "dynamov2_packages @ git+ssh://git@github.com/iambcl/dynamov2_packages.git@main",
    "langextract>=1.1.1",
    "plotly>=6.5.2",
    "pytest>=9.0.2",
    "openhands-sdk>=1.11.1",
    "openhands-tools>=1.11.1",
    "openhands-workspace>=1.11.1",
    "openhands-agent-server>=1.11.1",
]

[tool.setuptools]
py-modules = []
